Lots of cleanup, slight memory optimizations
A bit of code refactoring, made a few ranges more idiomatic so that they
performed faster.  Also restructured some constants in to a struct since
they were generally used together.  Given the unfortunate size of the
struct (33 bits), and the fact that the legal moves bool is generally
sufficient, might actually perform faster if that was pulled out and
left as its own slice.  Code runs 3% faster.  Whee.